crawler/requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0
supabase==2.1.0
schedule==1.2.0
fake-useragent==1.4.0
jieba==0.42.1
loguru==0.7.2
aiohttp==3.9.1